--- .~lock.data.csv#
@@ -1,1 +1,0 @@
-,DESKTOP-177OGLU/Luis,DESKTOP-177OGLU,20.05.2022 03:53,file:///C:/Users/Luis/AppData/Roaming/LibreOffice/4;
