--- Readme.txt
@@ -1,18 +1,0 @@
-Available languages:
--'ES' (spanish)
--'EN' (english)
-
-Available inputs:
-These are the ids of the tables you request, they can be found here for the spanish INE (Instituto Nacional de Estadística):
-
-https://www.ine.es/dyngs/INEbase/listaoperaciones.htm
-
-Optional nult:
-If not left blank, it will give the n last terms of the requested table
-
-Optional date:
-If not left blank, it will give the terms of the requested table in:
-
--date= YYYYMMDD (a given date)
--date= YYYYMMDD&date=YYYYMMDD (a list of given dates) 
--date=YYYYMMDD:YYYYMMDD (a range of dates)
