--- changelog
@@ -1,28 +1,0 @@
-Version 0.1
- * @Date 2022/4/21
- The program allows to create dynamically objects from json-stat files from the INE database.
- The URL to get the json-stat file is constructed from a config.yaml file.
- This are the classes of objects that can be created:
-  ProcJsonStatDataset contains:
-   -Dimensions (JsonStatDimension)
-   -Value list
-   -Status list
-  JsonStatDimension contains:
-    -Name
-    -Label
-    -Category (JsonStatCategory)
-    -Role
-  JsonStatCategory contains:
-    -Index
-    -Label
-    -Size
-  The names of the dimensions are normalized before being generated.
-
-Version 0.2
- * @Date 2022/4/28
- The program's methods are now inside the class IneJsonStat.
- The program now uses a log of two levels: INFO and DEBUG
- The URL is now checked only once only it's parameterized and fails, in which case
-  it is shortened to a basic URL and the system retries.
- Minor errors solved
-
