--- config.yaml
@@ -1,3 +1,0 @@
-- Details:
-    ttl: "W2"
-
