--- data_model/ine_status.py
@@ -1,8 +1,0 @@
-import enum
-
-
-class Status(enum.Enum):
-   D = 1
-   P = 2
-   E = 3
-   UNKNOWN = 0
