--- data_model/languages_enum.py
@@ -1,5 +1,0 @@
-from enum import Enum
-
-class LanguageEnum(Enum):
-    ES = 'ES'
-    EN = 'EN'
