+++ inejsonstat/config_files/config.yaml
@@ -1,0 +1,3 @@
+- Details:
+    ttl: "w1"
+
