+++ inejsonstat/dimension_enum_dc.py
@@ -1,0 +1,4 @@
+from collections import namedtuple
+
+DimensionItem = namedtuple("DimensionItem",  ["id", "name", "columns", "dataframe", "values_df", "status_df", "data",
+                                              "values", "status"])
