+++ inejsonstat/ine_status.py
@@ -1,0 +1,8 @@
+import enum
+
+
+class Status(enum.Enum):
+   D = "Definitive"
+   P = "Provisional"
+   E = "Estimated"
+   UNKNOWN = "Unknown"
