+++ inejsonstat/languages_enum.py
@@ -1,0 +1,5 @@
+from enum import Enum
+
+class LanguageEnum(Enum):
+    ES = 'ES'
+    EN = 'EN'
