+++ testing.py
@@ -1,29 +1,34 @@
 import datetime
 
-from data_model.ine_initializer import IneInitializer as ine_init
+import inejsonstat.__init__ as ine_init
 
 if __name__ == "__main__":
     date = datetime.date(year=2021, month=5, day=1)
     date2 = datetime.date(year=2021, month=4, day=1)
     date3 = datetime.date(year=2021, month=3, day=1)
+    date4 = datetime.date(year=2023, month=3, day=1)
 
     # Initialize the program
-    ine = ine_init.create2()
+    ine = ine_init.create()
 
     # Example with written date and language
     ine.do_request(target='2074', language='ES', date=date, datetype="range")
-    ine.do_request(target=ine.targets.N2065, language=ine.languages.EN, date="20210501")
+    print("From the url:", ine.last_url)
+    ine.do_request(target=ine.targets.N2065, language=ine.languages.EN, date="20210501:20210601")
+    print("From the url:", ine.last_url)
 
     # Example of ranged date
     ine.do_request(target=ine.targets.N2065, language=ine.languages.EN, date=[date, date2], datetype="range", nult=2)
+    print("From the url:", ine.last_url)
 
     # Main example
     # Get the json_stat data
     print("[1]-----------------------------------------------------------------------------")
-    json_data = ine.do_request(target=ine.targets.N2074, language=ine.languages.ES, date=[date, date2, date3],
-                               datetype="list", nult=123)
+    json_data = ine.do_request(target=ine.targets.N2074, language=ine.languages.ES, date=[date, date2,date3],
+                               datetype="list")
     print("The Json data is:")
     print(json_data)
+    print("From the url:", ine.last_url)
 
     # Initialize dataset
     print("[2]-----------------------------------------------------------------------------")
@@ -41,17 +46,21 @@ if __name__ == "__main__":
 
     print("[5]-----------------------------------------------------------------------------")
     print("Dataset dimensions are: ", dataset.dimensions)
+    print("Dimensión:", dataset.comunidadesautonomasyprovincias)
+    print("Dimensión>category>label:", dataset.comunidadesautonomasyprovincias.category.label)
+    dataset.comunidadesautonomasyprovincias.print_properties()
+    dataset.comunidadesautonomasyprovincias.category.print_properties()
 
     print("[6]-----------------------------------------------------------------------------")
     print("Dataset attributes: ")
     dataset.print_attributes()
 
     print("[7]-----------------------------------------------------------------------------")
-    print("Dataset value list:", dataset.value)
+    print("Dataset value list:", len(dataset.value))
     print("Size of value list: ", dataset.value_size)
 
     print("[8]-----------------------------------------------------------------------------")
-    print("Dataset status list:", dataset.status)
+    print("Dataset status list:", len(dataset.status))
     print("Size of status list: ", dataset.status_size)
 
     print("[9]-----------------------------------------------------------------------------")
@@ -97,5 +106,5 @@ if __name__ == "__main__":
     print("[18]-----------------------------------------------------------------------------")
     # There can be multiple values for a query
     df4 = ine.query(comunidadesautonomasyprovincias=[dataset.COMUNIDADESAUTONOMASYPROVINCIAS.BADAJOZ,
-                                                     dataset.COMUNIDADESAUTONOMASYPROVINCIAS.GRANADA], status="NO")
+                                                     "Granada"], status="NO")
     print(df4)
